Rename `datasets.h` to `table_config.h`
