+++ src/tests/test_cfd_relation_data.cpp
@@ -4,8 +4,8 @@
 #include <gtest/gtest.h>
 
 #include "algorithms/cfd/model/cfd_relation_data.h"
-#include "datasets.h"
 #include "parser/csv_parser/csv_parser.h"
+#include "table_config.h"
 
 namespace tests {
