+++ src/tests/test_typed_column_data.cpp
@@ -4,9 +4,9 @@
 #include <gtest/gtest.h>
 
 #include "algorithms/fd/fd_algorithm.h"
-#include "datasets.h"
 #include "model/table/column_layout_typed_relation_data.h"
 #include "parser/csv_parser/csv_parser.h"
+#include "table_config.h"
 
 namespace tests {
