+++ src/tests/test_util.cpp
@@ -4,12 +4,12 @@
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
-#include "datasets.h"
 #include "fd/pyro/model/list_agree_set_sample.h"
 #include "levenshtein_distance.h"
 #include "model/table/agree_set_factory.h"
 #include "model/table/column_layout_relation_data.h"
 #include "model/table/identifier_set.h"
+#include "table_config.h"
 
 namespace tests {
